Order of Variable Levels from Eurostat Dictionary.

Orders the factor levels.

dic_order(x, dic, type)

Arguments

x
a variable (code or labelled) to get order for.

dic
a name of the dictionary. Correspond a variable name in the data_frame from get_eurostat. Can be also data_frame from get_eurostat_dic.

type
a type of the x. Could be code or label.

Value

A numeric vector of orders.

Details

Some variables, like classifications, have logical or conventional ordering. Eurostat data tables are nor neccessary ordered in this order. The function dic_order get the ordering from Eurostat classifications dictionaries. The function label_eurostat can also order factor levels of labels with argument eu_order = TRUE.