Download Table of Contents of Eurostat Data Sets

Download table of contents (TOC) of eurostat datasets.

get_eurostat_toc()

Value

A tibble with eight columns:

title: The name of dataset of theme.

code: The codename of dataset of theme, will be used by the get_eurostat() and get_eurostat_raw() functions.

type: Is it a dataset, folder or table.

last.update.of.data, last.table.structure.change, data.start, data.end: Dates.

Details

The TOC is downloaded from https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?sort=1&file=table_of_contents_en.txt. The values in column 'code' should be used to download a selected dataset.

References

See citation("eurostat"):

# 
# Kindly cite the eurostat R package as follows:
# 
#   (C) Leo Lahti, Janne Huovari, Markus Kainu, Przemyslaw Biecek.
#   Retrieval and analysis of Eurostat open data with the eurostat
#   package. R Journal 9(1):385-392, 2017. Version 3.7.9 doi:
#   10.32614/RJ-2017-019 Package URL: http://ropengov.github.io/eurostat
#   Article URL:
#   https://journal.r-project.org/archive/2017/RJ-2017-019/index.html
# 
# A BibTeX entry for LaTeX users is
# 
#   @Article{,
#     title = {Retrieval and Analysis of Eurostat Open Data with the
#            eurostat Package},
#     author = {Leo Lahti and Janne Huovari and Markus Kainu and Przemyslaw Biecek},
#     journal = {The R Journal},
#     volume = {9},
#     number = {1},
#     pages = {385--392},
#     year = {2017},
#     doi = {10.32614/RJ-2017-019},
#     url = {https://doi.org/10.32614/RJ-2017-019},
#     note = {Version 3.7.9},
#   }

See also

Author

Przemyslaw Biecek and Leo Lahti ropengov-forum@googlegroups.com

Examples

# \donttest{
tmp <- get_eurostat_toc()
head(tmp)
#> # A tibble: 6 × 8
#>   title    code  type  `last update of… `last table str… `data start` `data end`
#>   <chr>    <chr> <chr> <chr>            <chr>            <chr>        <chr>     
#> 1 Databas… data  fold… NA               NA               NA           NA        
#> 2 General… gene… fold… NA               NA               NA           NA        
#> 3 Europea… euro… fold… NA               NA               NA           NA        
#> 4 Busines… ei_b… fold… NA               NA               NA           NA        
#> 5 Consume… ei_b… fold… NA               NA               NA           NA        
#> 6 Consume… ei_b… data… 29.09.2021       29.09.2021       1980M01      2021M09   
#> # … with 1 more variable: values <chr>
# }