Recode geo labels and rename regions from NUTS2013 to NUTS2016

Eurostat mixes NUTS2013 and NUTS2016 geographic label codes in the 'geo' column, which creates time-wise comparativity issues.

This function is deprecated, and a more general function was moved to regions::recode_nuts.

recode_to_nuts_2016(dat)

Arguments

dat
A Eurostat data frame downloaded with get_eurostat.

Value

An augmented and potentially relabelled data frame which contains all formerly 'NUTS2013' definition geo labels in the 'NUTS2016' vocabulary when only the code changed, but the boundary did not. It also contains some information on other geo labels that cannot be brought to the current 'NUTS2016' definition. Furthermore, when the official name of the region changed, it will use the new name (if the otherwise the region boundary did not change.) If not called before, the function will use the helper function harmonize_geo_code

Author

Daniel Antal

Examples

test_regional_codes <- data.frame ( 
  geo = c("FRB", "FRE", "UKN02", "IE022", "FR243", "FRB03"),
  time = c(rep(as.Date ("2014-01-01"), 5), as.Date("2015-01-01")), 
  values = c(1:6), 
  control = c("Changed from NUTS2 to NUTS1", 
              "New region NUTS2016 only", 
              "Discontinued region NUTS2013", 
              "Boundary shift NUTS2013", 
              "Recoded in NUTS2013", 
              "Recoded in NUTS2016"
  )) 
 
recode_to_nuts_2016(test_regional_codes)

#> Warning: The 'recode_to_nuts_2013' function is deprecated. Use instead regions::recode_nuts(dat, nuts_year = 2016)

#>     geo       time values                      control     typology
#> 1   FRB 2014-01-01      1  Changed from NUTS2 to NUTS1 nuts_level_1
#> 2   FRE 2014-01-01      2     New region NUTS2016 only nuts_level_1
#> 3 FRB03 2015-01-01      6          Recoded in NUTS2016 nuts_level_3
#> 4 IE022 2014-01-01      4      Boundary shift NUTS2013 nuts_level_3
#> 5 FR243 2014-01-01      5          Recoded in NUTS2013 nuts_level_3
#> 6 UKN02 2014-01-01      3 Discontinued region NUTS2013 nuts_level_3
#>                               typology_change code_2016
#> 1                                        <NA>       FRB
#> 2                                        <NA>       FRE
#> 3                                        <NA>     FRB03
#> 4 Recoded from IE022 [used in NUTS 2013-2013]     IE062
#> 5 Recoded from FR243 [used in NUTS 1999-2013]     FRB03
#> 6                      Used in NUTS 1999-2013      <NA>