Grep Datasets Titles from Eurostat

Lists names of dataset from eurostat with the particular pattern in the description.

search_eurostat(pattern, type = "dataset", fixed = TRUE)

grepEurostatTOC(pattern, type = "dataset")

Arguments

pattern

Character, datasets, folder or tables with this pattern in the description will be returned (depending on the 'type' argument)

type

Grep the Eurostat table of contents either for 'dataset' (default), 'folder', 'table' or "all" (for all types).

fixed

logical. If TRUE, pattern is a string to be matched as is. Change to FALSE if more complex regex matching is needed.

Value

A tibble with eight columns

titleThe name of dataset of theme

codeThe codename of dataset of theme, will be used by the get_eurostat and get_eurostat_raw functions.

typeIs it a dataset, folder or table.

last.update.of.data, last.table.structure.change, data.start, data.endDates.

Details

Downloads list of all datasets available on eurostat and return list of names of datasets that contains particular pattern in the dataset description. E.g. all datasets related to education of teaching.

Functions

grepEurostatTOC: Old deprecated version

References

See citation("eurostat")

See also

get_eurostat, get_eurostat_toc

Examples

## Not run: ------------------------------------
#   tmp <- search_eurostat("education")
#   head(tmp)
#   # Use "fixed = TRUE" when pattern has characters that would need escaping.
#   # Here, parentheses would normally need to be escaped in regex
#   tmp <- search_eurostat("Live births (total) by NUTS 3 region", fixed = TRUE)
## ---------------------------------------------