Cuts the Values Column into Classes and Polishes the Labels

Categorises a numeric vector into automatic or manually defined categories and polishes the labels ready for used in mapping with ggplot2.

Usage

cut_to_classes(
  x,
  n = 5,
  style = "equal",
  manual = FALSE,
  manual_breaks = NULL,
  decimals = 0,
  nodata_label = "No data"
)

Arguments

x

A numeric vector, eg. values variable in data returned by get_eurostat().

n

A numeric. number of classes/categories

style

chosen style: one of "fixed", "sd", "equal", "pretty", "quantile", "kmeans", "hclust", "bclust", "fisher", "jenks", "dpih", "headtails", "maximum", or "box"

manual

Logical. If manual breaks are being used

manual_breaks

Numeric vector with manual threshold values

decimals

Number of decimals to include with labels

nodata_label

String. Text label for NA category.

Value

a factor.

See also

Author

Markus Kainu markuskainu@gmail.com

Examples

# \donttest{
# lp <- get_eurostat("nama_aux_lp")
lp <- get_eurostat("nama_10_lp_ulc")
#> Table nama_10_lp_ulc cached at /tmp/RtmpoLH6G0/eurostat/33173e1cf98fd4be6e937231eca9f7f1.rds
lp$class <- cut_to_classes(lp$values, n = 5, style = "equal", decimals = 1)
#> Warning: var has missing values, omitted in finding classes
#> Warning: var has missing values, omitted in finding classes
# }