R Tools for Eurostat open data

Brief summary of the eurostat package

Details

Package	eurostat

Type	Package

Version	3.7.9

Date	`2014-2021

License	BSD_2_clause + file LICENSE

LazyLoad	yes

R Tools for Eurostat Open Data

regions functions

For working with sub-national statistics the basic functions of the regions package are imported https://regions.dataobservatory.eu/.

References

See citation("eurostat"):

# 
# Kindly cite the eurostat R package as
# follows:
# 
#   (C) Leo Lahti, Janne Huovari, Markus
#   Kainu, Przemyslaw Biecek. Retrieval and
#   analysis of Eurostat open data with the
#   eurostat package. R Journal 9(1):385-392,
#   2017. Version 3.7.9 doi:
#   10.32614/RJ-2017-019 Package URL:
#   http://ropengov.github.io/eurostat Article
#   URL:
#   https://journal.r-project.org/archive/2017/RJ-2017-019/index.html
# 
# A BibTeX entry for LaTeX users is
# 
#   @Article{,
#     title = {Retrieval and Analysis of Eurostat Open Data with the
#            eurostat Package},
#     author = {Leo Lahti and Janne Huovari and Markus Kainu and Przemyslaw Biecek},
#     journal = {The R Journal},
#     volume = {9},
#     number = {1},
#     pages = {385--392},
#     year = {2017},
#     doi = {10.32614/RJ-2017-019},
#     url = {https://doi.org/10.32614/RJ-2017-019},
#     note = {Version 3.7.9},
#   }

See also

Author

Leo Lahti, Janne Huovari, Markus Kainu, Przemyslaw Biecek

Examples

library(eurostat)