Date Conversion from Eurostat Time Format

Date conversion from Eurostat time format. A function to convert Eurostat time values to objects of class Date() representing calendar dates.

Usage

eurotime2date(x, last = FALSE)

Arguments

x

a charter string with time information in Eurostat time format.

last

a logical. If FALSE (default) the date is the first date of the period (month, quarter or year). If TRUE the date is the last date of the period.

Value

an object of class Date().

References

See citation("eurostat"):

# 
# Kindly cite the eurostat R package as follows:
# 
#   (C) Leo Lahti, Janne Huovari, Markus Kainu, Przemyslaw Biecek.
#   Retrieval and analysis of Eurostat open data with the eurostat
#   package. R Journal 9(1):385-392, 2017. doi: 10.32614/RJ-2017-019
#   Package URL: http://ropengov.github.io/eurostat Article URL:
#   https://journal.r-project.org/archive/2017/RJ-2017-019/index.html
# 
# A BibTeX entry for LaTeX users is
# 
#   @Article{,
#     title = {Retrieval and Analysis of Eurostat Open Data with the eurostat Package},
#     author = {Leo Lahti and Janne Huovari and Markus Kainu and Przemyslaw Biecek},
#     journal = {The R Journal},
#     volume = {9},
#     number = {1},
#     pages = {385--392},
#     year = {2017},
#     doi = {10.32614/RJ-2017-019},
#     url = {https://doi.org/10.32614/RJ-2017-019},
#   }

See also

Author

Janne Huovari janne.huovari@ptt.fi

Examples

# \donttest{
na_q <- get_eurostat("namq_10_pc", time_format = "raw")
#> Table namq_10_pc cached at /tmp/RtmpBLpMpp/eurostat/namq_10_pc_raw_code_FF.rds
na_q$time <- eurotime2date(x = na_q$time)
unique(na_q$time)
#>   [1] "2021-10-01" "2021-07-01" "2021-04-01" "2021-01-01" "2020-10-01"
#>   [6] "2020-07-01" "2020-04-01" "2020-01-01" "2019-10-01" "2019-07-01"
#>  [11] "2019-04-01" "2019-01-01" "2018-10-01" "2018-07-01" "2018-04-01"
#>  [16] "2018-01-01" "2017-10-01" "2017-07-01" "2017-04-01" "2017-01-01"
#>  [21] "2016-10-01" "2016-07-01" "2016-04-01" "2016-01-01" "2015-10-01"
#>  [26] "2015-07-01" "2015-04-01" "2015-01-01" "2014-10-01" "2014-07-01"
#>  [31] "2014-04-01" "2014-01-01" "2013-10-01" "2013-07-01" "2013-04-01"
#>  [36] "2013-01-01" "2012-10-01" "2012-07-01" "2012-04-01" "2012-01-01"
#>  [41] "2011-10-01" "2011-07-01" "2011-04-01" "2011-01-01" "2010-10-01"
#>  [46] "2010-07-01" "2010-04-01" "2010-01-01" "2009-10-01" "2009-07-01"
#>  [51] "2009-04-01" "2009-01-01" "2008-10-01" "2008-07-01" "2008-04-01"
#>  [56] "2008-01-01" "2007-10-01" "2007-07-01" "2007-04-01" "2007-01-01"
#>  [61] "2006-10-01" "2006-07-01" "2006-04-01" "2006-01-01" "2005-10-01"
#>  [66] "2005-07-01" "2005-04-01" "2005-01-01" "2004-10-01" "2004-07-01"
#>  [71] "2004-04-01" "2004-01-01" "2003-10-01" "2003-07-01" "2003-04-01"
#>  [76] "2003-01-01" "2002-10-01" "2002-07-01" "2002-04-01" "2002-01-01"
#>  [81] "2001-10-01" "2001-07-01" "2001-04-01" "2001-01-01" "2000-10-01"
#>  [86] "2000-07-01" "2000-04-01" "2000-01-01" "1999-10-01" "1999-07-01"
#>  [91] "1999-04-01" "1999-01-01" "1998-10-01" "1998-07-01" "1998-04-01"
#>  [96] "1998-01-01" "1997-10-01" "1997-07-01" "1997-04-01" "1997-01-01"
#> [101] "1996-10-01" "1996-07-01" "1996-04-01" "1996-01-01" "1995-10-01"
#> [106] "1995-07-01" "1995-04-01" "1995-01-01" "1994-10-01" "1994-07-01"
#> [111] "1994-04-01" "1994-01-01" "1993-10-01" "1993-07-01" "1993-04-01"
#> [116] "1993-01-01" "1992-10-01" "1992-07-01" "1992-04-01" "1992-01-01"
#> [121] "1991-10-01" "1991-07-01" "1991-04-01" "1991-01-01" "1990-10-01"
#> [126] "1990-07-01" "1990-04-01" "1990-01-01" "1989-10-01" "1989-07-01"
#> [131] "1989-04-01" "1989-01-01" "1988-10-01" "1988-07-01" "1988-04-01"
#> [136] "1988-01-01" "1987-10-01" "1987-07-01" "1987-04-01" "1987-01-01"
#> [141] "1986-10-01" "1986-07-01" "1986-04-01" "1986-01-01" "1985-10-01"
#> [146] "1985-07-01" "1985-04-01" "1985-01-01" "1984-10-01" "1984-07-01"
#> [151] "1984-04-01" "1984-01-01" "1983-10-01" "1983-07-01" "1983-04-01"
#> [156] "1983-01-01" "1982-10-01" "1982-07-01" "1982-04-01" "1982-01-01"
#> [161] "1981-10-01" "1981-07-01" "1981-04-01" "1981-01-01" "1980-10-01"
#> [166] "1980-07-01" "1980-04-01" "1980-01-01"
# }