Conversion of Eurostat Time Format to Numeric

A conversion of a Eurostat time format to numeric.

Usage

eurotime2num(x)

Arguments

x

a charter string with time information in Eurostat time format.

Value

see as.numeric().

Details

Bi-annual, quarterly and monthly data is presented as fraction of the year in beginning of the period. Conversion of daily data is not supported.

See also

Author

Janne Huovari janne.huovari@ptt.fi

Examples

# \donttest{
na_q <- get_eurostat("namq_10_pc", time_format = "raw")
na_q$time <- eurotime2num(x = na_q$time)

unique(na_q$time)
#>   [1] 2022.75 2022.50 2022.25 2022.00 2021.75 2021.50 2021.25 2021.00 2020.75
#>  [10] 2020.50 2020.25 2020.00 2019.75 2019.50 2019.25 2019.00 2018.75 2018.50
#>  [19] 2018.25 2018.00 2017.75 2017.50 2017.25 2017.00 2016.75 2016.50 2016.25
#>  [28] 2016.00 2015.75 2015.50 2015.25 2015.00 2014.75 2014.50 2014.25 2014.00
#>  [37] 2013.75 2013.50 2013.25 2013.00 2012.75 2012.50 2012.25 2012.00 2011.75
#>  [46] 2011.50 2011.25 2011.00 2010.75 2010.50 2010.25 2010.00 2009.75 2009.50
#>  [55] 2009.25 2009.00 2008.75 2008.50 2008.25 2008.00 2007.75 2007.50 2007.25
#>  [64] 2007.00 2006.75 2006.50 2006.25 2006.00 2005.75 2005.50 2005.25 2005.00
#>  [73] 2004.75 2004.50 2004.25 2004.00 2003.75 2003.50 2003.25 2003.00 2002.75
#>  [82] 2002.50 2002.25 2002.00 2001.75 2001.50 2001.25 2001.00 2000.75 2000.50
#>  [91] 2000.25 2000.00 1999.75 1999.50 1999.25 1999.00 1998.75 1998.50 1998.25
#> [100] 1998.00 1997.75 1997.50 1997.25 1997.00 1996.75 1996.50 1996.25 1996.00
#> [109] 1995.75 1995.50 1995.25 1995.00 1994.75 1994.50 1994.25 1994.00 1993.75
#> [118] 1993.50 1993.25 1993.00 1992.75 1992.50 1992.25 1992.00 1991.75 1991.50
#> [127] 1991.25 1991.00 1990.75 1990.50 1990.25 1990.00 1989.75 1989.50 1989.25
#> [136] 1989.00 1988.75 1988.50 1988.25 1988.00 1987.75 1987.50 1987.25 1987.00
#> [145] 1986.75 1986.50 1986.25 1986.00 1985.75 1985.50 1985.25 1985.00 1984.75
#> [154] 1984.50 1984.25 1984.00 1983.75 1983.50 1983.25 1983.00 1982.75 1982.50
#> [163] 1982.25 1982.00 1981.75 1981.50 1981.25 1981.00 1980.75 1980.50 1980.25
#> [172] 1980.00
# }