Download Eurostat Dictionary

Download a Eurostat dictionary.

Usage

get_eurostat_dic(dictname, lang = "en")

Arguments

dictname

A character, dictionary for the variable to be downloaded.

lang

A character, language code. Options: "en" (default), "fr", "de".

Value

tibble with two columns: code names and full names.

Details

For given coded variable from Eurostat https://ec.europa.eu/eurostat/. The dictionaries link codes with human-readable labels. To translate codes to labels, use label_eurostat().

References

See citation("eurostat"):

# 
# Kindly cite the eurostat R package as follows:
# 
#   (C) Leo Lahti, Janne Huovari, Markus Kainu, Przemyslaw Biecek.
#   Retrieval and analysis of Eurostat open data with the eurostat
#   package. R Journal 9(1):385-392, 2017. doi: 10.32614/RJ-2017-019
#   Package URL: http://ropengov.github.io/eurostat Article URL:
#   https://journal.r-project.org/archive/2017/RJ-2017-019/index.html
# 
# A BibTeX entry for LaTeX users is
# 
#   @Article{,
#     title = {Retrieval and Analysis of Eurostat Open Data with the eurostat Package},
#     author = {Leo Lahti and Janne Huovari and Markus Kainu and Przemyslaw Biecek},
#     journal = {The R Journal},
#     volume = {9},
#     number = {1},
#     pages = {385--392},
#     year = {2017},
#     doi = {10.32614/RJ-2017-019},
#     url = {https://doi.org/10.32614/RJ-2017-019},
#   }

See also

Author

Przemyslaw Biecek and Leo Lahti leo.lahti@iki.fi. Thanks to Wietse Dol for contributions.

Examples

# \donttest{
get_eurostat_dic("crop_pro")
#> # A tibble: 224 × 2
#>    code_name full_name                                                    
#>    <chr>     <chr>                                                        
#>  1 C1040     Cereals for the production of grain (including rice and seed)
#>  2 C1050     Cereals (excluding rice)                                     
#>  3 C1100     Wheat (including spelt)                                      
#>  4 C1120     Common wheat and spelt                                       
#>  5 C1123     Common winter wheat                                          
#>  6 C1124     Common spring wheat                                          
#>  7 C1130     Durum wheat                                                  
#>  8 C1133     Winter durum wheat                                           
#>  9 C1134     Spring durum wheat                                           
#> 10 C1140     Rye and maslin                                               
#> # … with 214 more rows
#> # ℹ Use `print(n = ...)` to see more rows

# Try another language
get_eurostat_dic("crop_pro", lang = "fr")
#> # A tibble: 224 × 2
#>    code_name full_name                                                     
#>    <chr>     <chr>                                                         
#>  1 C1040     Céréales pour la production de grains (riz et semence compris)
#>  2 C1050     Céréales (à l'exception du riz)                               
#>  3 C1100     Blé (épeautre compris)                                        
#>  4 C1120     Blé tendre et épeautre                                        
#>  5 C1123     Blé tendre d'hiver                                            
#>  6 C1124     Blé tendre de printemps                                       
#>  7 C1130     Blé dur                                                       
#>  8 C1133     Blé dur d'hiver                                               
#>  9 C1134     Blé dur de printemps                                          
#> 10 C1140     Seigle et méteil                                              
#> # … with 214 more rows
#> # ℹ Use `print(n = ...)` to see more rows
# }