Download Geospatial Data from GISGO

Downloads either a simple features (sf), SpatialPolygonDataFrame or a data_frame preprocessed using broom::tidy().

get_eurostat_geospatial(output_class = "sf", resolution = "60",
  nuts_level = "all", year = "2016", cache = TRUE,
  update_cache = FALSE, cache_dir = NULL)

Arguments

output_class
A string. Class of object returned, either sf simple features, df (data_frame) or spdf (SpatialPolygonDataFrame)

resolution
Resolution of the geospatial data. One of "60" (1:60million), "20" (1:20million) "10" (1:10million) "03" (1:3million) or "01" (1:1million).

nuts_level
Level of NUTS classification of the geospatial data. One of "0", "1", "2", "3" or "all" (mimics the original behaviour)

year
NUTS release year. One of "2003", "2006", "2010", "2013" or "2016"

cache
a logical whether to do caching. Default is TRUE. Affects only queries from the bulk download facility.

update_cache
a locigal whether to update cache. Can be set also with options(eurostat_update = TRUE)

cache_dir
a path to a cache directory. The directory have to exist. The NULL (default) uses and creates 'eurostat' directory in the temporary directory from tempdir. Directory can also be set with option eurostat_cache_dir.

Value

a sf, data_frame or SpatialPolygonDataFrame.

Details

The data source URL is http://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units.

Examples

sf <- get_eurostat_geospatial(output_class = "sf", resolution = "60", nuts_level = "all")

#> 
#> COPYRIGHT NOTICE
#> 
#> When data downloaded from this page 
#> <http://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units>
#> is used in any printed or electronic publication, 
#> in addition to any other provisions 
#> applicable to the whole Eurostat website, 
#> data source will have to be acknowledged 
#> in the legend of the map and 
#> in the introductory page of the publication 
#> with the following copyright notice:
#> 
#> - EN: (C) EuroGeographics for the administrative boundaries
#> - FR: (C) EuroGeographics pour les limites administratives
#> - DE: (C) EuroGeographics bezuglich der Verwaltungsgrenzen
#> 
#> For publications in languages other than 
#> English, French or German, 
#> the translation of the copyright notice 
#> in the language of the publication shall be used.
#> 
#> If you intend to use the data commercially, 
#> please contact EuroGeographics for 
#> information regarding their licence agreements.
#>           

#> sf at resolution 1:60 read from local file

#> 
#> # --------------------------
#> HEADS UP!!
#> 
#> Function now returns the data in 'sf'-class (simple features) 
#> by default which is different 
#> from previous behaviour's 'SpatialPolygonDataFrame'. 
#> 
#> If you prefer either 'SpatialPolygonDataFrame' or 
#> fortified 'data_frame' (for ggplot2::geom_polygon), 
#> please specify it explicitly to 'output_class'-argument!
#> 
#> # --------------------------          
#>           

df <- get_eurostat_geospatial(output_class = "df", resolution = "20", nuts_level = "0")

#> 
#> COPYRIGHT NOTICE
#> 
#> When data downloaded from this page 
#> <http://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units>
#> is used in any printed or electronic publication, 
#> in addition to any other provisions 
#> applicable to the whole Eurostat website, 
#> data source will have to be acknowledged 
#> in the legend of the map and 
#> in the introductory page of the publication 
#> with the following copyright notice:
#> 
#> - EN: (C) EuroGeographics for the administrative boundaries
#> - FR: (C) EuroGeographics pour les limites administratives
#> - DE: (C) EuroGeographics bezuglich der Verwaltungsgrenzen
#> 
#> For publications in languages other than 
#> English, French or German, 
#> the translation of the copyright notice 
#> in the language of the publication shall be used.
#> 
#> If you intend to use the data commercially, 
#> please contact EuroGeographics for 
#> information regarding their licence agreements.
#>           

#> No encoding supplied: defaulting to UTF-8.

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Regions defined for each Polygons

#> Joining, by = "id"

#> 
#> # --------------------------
#> HEADS UP!!
#> 
#> Function now returns the data in 'sf'-class (simple features) 
#> by default which is different 
#> from previous behaviour's 'SpatialPolygonDataFrame'. 
#> 
#> If you prefer either 'SpatialPolygonDataFrame' or 
#> fortified 'data_frame' (for ggplot2::geom_polygon), 
#> please specify it explicitly to 'output_class'-argument!
#> 
#> # --------------------------          
#>           

spdf <- get_eurostat_geospatial(output_class = "spdf", resolution = "10", nuts_level = "3")

#> 
#> COPYRIGHT NOTICE
#> 
#> When data downloaded from this page 
#> <http://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units>
#> is used in any printed or electronic publication, 
#> in addition to any other provisions 
#> applicable to the whole Eurostat website, 
#> data source will have to be acknowledged 
#> in the legend of the map and 
#> in the introductory page of the publication 
#> with the following copyright notice:
#> 
#> - EN: (C) EuroGeographics for the administrative boundaries
#> - FR: (C) EuroGeographics pour les limites administratives
#> - DE: (C) EuroGeographics bezuglich der Verwaltungsgrenzen
#> 
#> For publications in languages other than 
#> English, French or German, 
#> the translation of the copyright notice 
#> in the language of the publication shall be used.
#> 
#> If you intend to use the data commercially, 
#> please contact EuroGeographics for 
#> information regarding their licence agreements.
#>           

#> No encoding supplied: defaulting to UTF-8.

#> 
#> # --------------------------
#> HEADS UP!!
#> 
#> Function now returns the data in 'sf'-class (simple features) 
#> by default which is different 
#> from previous behaviour's 'SpatialPolygonDataFrame'. 
#> 
#> If you prefer either 'SpatialPolygonDataFrame' or 
#> fortified 'data_frame' (for ggplot2::geom_polygon), 
#> please specify it explicitly to 'output_class'-argument!
#> 
#> # --------------------------          
#>