Download Data from Eurostat Database

Download data from the eurostat database.

Usage

get_eurostat_raw(id)

Arguments

id

A code name for the dataset of interested. See the table of contents of eurostat datasets for more details.

Value

A dataset in tibble format. First column contains comma separated codes of cases. Other columns usually corresponds to years and column names are years with preceding X. Data is in character format as it contains values together with eurostat flags for data.

Details

Data is downloaded from https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing and transformed into tabular format.

References

See citation("eurostat"):

# 
# Kindly cite the eurostat R package as follows:
# 
#   (C) Leo Lahti, Janne Huovari, Markus Kainu, Przemyslaw Biecek.
#   Retrieval and analysis of Eurostat open data with the eurostat
#   package. R Journal 9(1):385-392, 2017. doi: 10.32614/RJ-2017-019
#   Package URL: http://ropengov.github.io/eurostat Article URL:
#   https://journal.r-project.org/archive/2017/RJ-2017-019/index.html
# 
# A BibTeX entry for LaTeX users is
# 
#   @Article{,
#     title = {Retrieval and Analysis of Eurostat Open Data with the eurostat Package},
#     author = {Leo Lahti and Janne Huovari and Markus Kainu and Przemyslaw Biecek},
#     journal = {The R Journal},
#     volume = {9},
#     number = {1},
#     pages = {385--392},
#     year = {2017},
#     doi = {10.32614/RJ-2017-019},
#     url = {https://doi.org/10.32614/RJ-2017-019},
#   }

See also

Author

Przemyslaw Biecek, Leo Lahti and Janne Huovari

Examples

# \donttest{
eurostat:::get_eurostat_raw("educ_iste")
#> # A tibble: 213 × 16
#>    `indic_ed,geo\\time` `2012` `2011` `2010` `2009` `2008` `2007` `2006` `2005`
#>    <chr>                <chr>  <chr>  <chr>  <chr>  <chr>  <chr>  <chr>  <chr> 
#>  1 ST1_1,AL             NA     NA     NA     NA     NA     NA     NA     NA    
#>  2 ST1_1,AT             10.1   10.2   10.4   10.6   11.0   11.5   11.7   11.8  
#>  3 ST1_1,BE             10.5 d 10.5 d 10.5 d 10.5 d 10.8 d 10.8 d 10.9 d 10.8 d
#>  4 ST1_1,BE_FRA         NA     10.1   10.2   10.3   NA     10.4   10.4   NA    
#>  5 ST1_1,BE_VLA         NA     10.8   10.7   10.7   11.0   11.1   11.2   NA    
#>  6 ST1_1,BG             13.9   13.8   13.6   13.5   12.8   12.8   12.9   13.2  
#>  7 ST1_1,CY             11.5   11.4   11.5   11.8   12.3   13.0   14.0   14.1  
#>  8 ST1_1,CZ             13.2   13.3   14.2 d 14.2 d 14.2 d 14.5 d 13.4   14.4  
#>  9 ST1_1,DE             15.4   15.7   16.1   16.6   16.7   16.9   17.2   17.2  
#> 10 ST1_1,DK             : u    : u    : u    : u    : u    : u    : u    : u   
#> # ℹ 203 more rows
#> # ℹ 7 more variables: `2004` <chr>, `2003` <chr>, `2002` <chr>, `2001` <chr>,
#> #   `2000` <chr>, `1999` <chr>, `1998` <chr>
# }