Download Data from Eurostat Database

Download data from the eurostat database.

get_eurostat_raw(id)

Arguments

id
A code name for the dataset of interested. See the table of contents of eurostat datasets for more details.

Value

A dataset in tibble format. First column contains comma separated codes of cases. Other columns usually corresponds to years and column names are years with preceding X. Data is in character format as it contains values together with eurostat flags for data.

Details

Data is downloaded from http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing and transformed into tabular format.

References

see citation("eurostat")

See also

Examples

# NOT RUN {
               tmp <- eurostat:::get_eurostat_raw("educ_iste")
               head(tmp)

# }