Harmonize NUTS region codes that changed with the NUTS2016 definition

Eurostat mixes NUTS2013 and NUTS2016 geographic label codes in the 'geo' column, which creates time-wise comparativity issues. This deprecated function checked if you data is affected by this problem and gives information on what to do.

This function is deprecated, and a more general function was moved to regions::validate_nuts_regions().

Usage

harmonize_geo_code(dat)

Arguments

dat

A Eurostat data frame downloaded with get_eurostat()

Value

An augmented data frame that explains potential problems and possible solutions.

See also

Author

Daniel Antal

Examples

dat <- eurostat::tgs00026
regions::validate_nuts_regions(dat)
#> # A tibble: 3,096 × 7
#>    unit     na_item geo   time  values typology     valid_2016
#>    <chr>    <chr>   <chr> <chr>  <dbl> <chr>        <lgl>     
#>  1 PPCS_HAB B6N     AT11  2005   18100 nuts_level_2 TRUE      
#>  2 PPCS_HAB B6N     AT12  2005   18900 nuts_level_2 TRUE      
#>  3 PPCS_HAB B6N     AT13  2005   19400 nuts_level_2 TRUE      
#>  4 PPCS_HAB B6N     AT21  2005   17800 nuts_level_2 TRUE      
#>  5 PPCS_HAB B6N     AT22  2005   17800 nuts_level_2 TRUE      
#>  6 PPCS_HAB B6N     AT31  2005   18400 nuts_level_2 TRUE      
#>  7 PPCS_HAB B6N     AT32  2005   18600 nuts_level_2 TRUE      
#>  8 PPCS_HAB B6N     AT33  2005   18000 nuts_level_2 TRUE      
#>  9 PPCS_HAB B6N     AT34  2005   18900 nuts_level_2 TRUE      
#> 10 PPCS_HAB B6N     BE10  2005   14800 nuts_level_2 TRUE      
#> # … with 3,086 more rows