Get Eurostat Codes

Get definitions for Eurostat codes from Eurostat dictionaries.

label_eurostat(x, dic = NULL, code = NULL, eu_order = FALSE,
  lang = "en", countrycode = NULL, countrycode_nomatch = NULL,
  custom_dic = NULL, fix_duplicated = FALSE)

label_eurostat_vars(x, lang = "en")

label_eurostat_tables(x, lang = "en")

Arguments

x
A character or a factor vector or a data_frame.

dic
A string (vector) naming eurostat dictionary or dictionaries. If NULL (default) dictionry names taken from column names of the data_frame.

code
For data_frames names of the column for which also code columns should be retained. The suffix "_code" is added to code column names.

eu_order
Logical. Should Eurostat ordering used for label levels. Affects only factors.

lang
A character, code for language. Available are "en" (default), "fr" and "de".

countrycode
A NULL or a name of the coding scheme for the countrycode to label "geo" variable with countrycode-package. It can be used to convert to short and long country names in many different languages. If NULL (default) eurostat dictionary is used instead.

countrycode_nomatch
What to do when using the countrycode to label a "geo" and countrycode fails to find a match, for example other than country codes like EU28. the original code is used with a NULL (default), eurostat dictionary label is used with "eurostat", and NA is used with NA.

custom_dic
a named vector or named list of named vectors to give an own dictionary for (part of) codes. Names of the vector should be codes and values labels. List can be used to spesify dictonaries and then list names should be dictionary codes.

fix_duplicated
A logical. If TRUE, the code is added to the duplicated label values. If FALSE (default) error is given if labelling produce duplicates.

Value

a vector or a data_frame.

Details

A character or a factor vector of codes returns a corresponding vector of definitions. label_eurostat labels also data_frames from get_eurostat. For vectors a dictionary name have to be supplied. For data_frames dictonary names are taken from column names. "time" and "values" columns are returned as they were, so you can supply data_frame from get_eurostat and get data_frame with definitions instead of codes.

Some Eurostat dictionaries includes dublicated labels. By default dublicated labels cause an error, but they can be fixed automatically with fix_duplicated = TRUE.

Functions

label_eurostat_vars: Get definitions for variable (column) names. For objects other than characters or factors definitions are get for names.

label_eurostat_tables: Get definitions for table names

Examples

 
# NOT RUN {
   lp <- get_eurostat("nama_10_lp_ulc")
   lpl <- label_eurostat(lp)
   str(lpl)
   lpl_order <- label_eurostat(lp, eu_order = TRUE)
   lpl_code <- label_eurostat(lp, code = "unit")
   label_eurostat_vars(names(lp))
   label_eurostat_tables("nama_10_lp_ulc")
   label_eurostat(c("FI", "DE", "EU28"), dic = "geo")
   label_eurostat(c("FI", "DE", "EU28"), dic = "geo", custom_dic = c(DE = "Germany"))
   label_eurostat(c("FI", "DE", "EU28"), dic = "geo", countrycode = "country.name",
                  custom_dic = c(EU28 = "EU"))
   label_eurostat(c("FI", "DE", "EU28"), dic = "geo", countrycode = "country.name")
   # In Finnish
   label_eurostat(c("FI", "DE", "EU28"), dic = "geo", countrycode = "cldr.short.fi")

# }