Transform Data into Row-Column-Value Format

Transform raw Eurostat data table into the row-column-value format (RCV).

tidy_eurostat(
  dat,
  time_format = "date",
  select_time = NULL,
  stringsAsFactors = FALSE,
  keepFlags = FALSE
)

Arguments

dat

a data_frame from get_eurostat_raw.

time_format

a string giving a type of the conversion of the time column from the eurostat format. A "date" (default) convers to a Date with a first date of the period. A "date_last" convers to a Date with a last date of the period. A "num" convers to a numeric and "raw" does not do conversion. See eurotime2date and eurotime2num.

select_time

a character symbol for a time frequence or NULL (default).

stringsAsFactors

if TRUE (the default) variables are converted to factors in original Eurostat order. If FALSE they are returned as strings.

keepFlags

a logical whether the flags (e.g. "confidential", "provisional") should be kept in a separate column or if they can be removed. Default is FALSE

Value

tibble in the molten format with the last column 'values'.

References

See citation("eurostat").

See also

Author

Przemyslaw Biecek, Leo Lahti and Janne Huovari